Generalied link-dirs script and fixed recursive link problem.
The old link-dirs script looked specifically for files starting '%3a' and
created symbol links for them, which also led it to build recursive symbolic
links when the original symbolic link to a directory already existed.  It also
did not handle files like the and-let%2a.sls file which need to be
and-let*.sls.  The new script fixes this, and only creates links after
verifying the link does not already exist, so it is safe to run multiple times,
or to run as new libraries are added.
  link-dirs.chezscheme.sps

Added vim swap files to the gitignore.
  .gitignore