Removed SRFI-32-error->R6RS call.
This error "translation" was producing an error out of context in the expander.
I'm not exactly sure what the issue with it is, but I have resolved it by just
having the code all the srfi 23 error instead of bothering with the r6rs error.
(Frankly, in an r6rs library, the srfi 23 error should just be calling the r6rs
error anyway.)
  %3a67/compare-procedures.sls