Use a free software implementation of SRFI 5 and add tests
The SRFI 5 code we were using was from the SRFI 5 document, which
carries a license that explicitly does not permit modifications.
Attempts to contact the SRFI author (Andrew Gaynor, AKA Silver) have
not been sucessful.

This commit replaces that implementation with a fresh one written by
myself and licensed under the MIT license. A test suite has also been
added.