Add SRFIs 165 and 166 (monadic formatting)
Also patches 130 to support string-copy with optional arguments, as in
the spec.

The SRFI 166 test suite is included but it depends on Alex Shinn’s
`(chibi test)`. This is available for Chez here:
<https://codeberg.org/dpk/chibi-lib>
but as it’s not (yet) a SRFI of its own, I’ve excluded it from here
and left the tests in the SRFI directory rather than putting them in
tests/. For me on Chez 10.0.0, all of the tests pass except for two
spurious failures (the test expects lowercase hex digits where Chez
gives uppercase ones, but this is not actually specified by the SRFI
so either is okay).